requirements.txt
# AI and Machine Learning Dependencies
anthropic>=0.25.0
openai>=1.0.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0

# Data Processing and OCR
pandas>=2.1.0
numpy<2.0.0,>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.0
opencv-python-headless==4.8.1.78
pdfplumber>=0.10.0
tabula-py>=2.9.0

# Database and APIs
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
redis>=5.0.0
requests>=2.31.0
aiohttp>=3.9.0
cachetools>=5.3.0
tenacity>=8.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
slowapi>=0.1.9
structlog>=23.2.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Document Processing
spacy>=3.7.0
nltk>=3.8.0
textstat>=0.7.0
